+++ Catalog.Web/Configurations/PipelineConfiguration.cs
@@ -1,0 +1,27 @@
+﻿using Catalog.Web.Middleware;
+
+namespace Catalog.Web.Configurations
+{
+    public static class PipelineConfiguration
+    {
+        public static WebApplication ConfigurePipeline(this WebApplication app)
+        {
+            if (app.Environment.IsDevelopment())
+            {
+                app.UseSwagger();
+                app.UseSwaggerUI();
+            }
+
+            app.UseMiddleware<ErrorHandlerMiddleware>();
+
+            app.UseHttpsRedirection();
+
+            app.UseAuthentication();
+            app.UseAuthorization();
+
+            app.MapControllers();
+
+            return app;
+        }
+    }
+}
