+++ Identity.Web/Configurations/PipelineConfiguration.cs
@@ -1,0 +1,26 @@
+﻿using Identity.Web.Middleware;
+
+namespace Identity.Web.Configurations
+{
+    public static class PipelineConfiguration
+    {
+        public static WebApplication ConfigurePipeline(this WebApplication app)
+        {
+            if (app.Environment.IsDevelopment())
+            {
+                app.UseSwagger();
+                app.UseSwaggerUI();
+            }
+
+            app.UseMiddleware<ErrorHandlerMiddleware>();
+
+            app.UseHttpsRedirection();
+
+            app.UseAuthorization();
+
+            app.MapControllers();
+
+            return app;
+        }
+    }
+}
