+++ Identity.Web/Program.cs
@@ -1,6 +1,8 @@
 using Identity.Infrastructure.Data;
-using Identity.Web;
+using Identity.Web.Configurations;
 using Serilog;
+
+SerilogConfiguration.ConfigureLogging();
 
 Log.Information("Starting up");
 
@@ -8,10 +10,7 @@ try
 {
     var builder = WebApplication.CreateBuilder(args);
 
-    builder.Host.UseSerilog((context, configuration) =>
-    {
-        configuration.ReadFrom.Configuration(context.Configuration);
-    });
+    builder.Host.UseSerilog();
 
     var app = builder
         .ConfigureServices()
