+++ Identity.Web/appsettings.json
@@ -20,17 +20,15 @@
   },
   "AllowedHosts": "*",
   "Serilog": {
-    "MinimumLevel": "Information",
-    "Override": {
-      "Microsoft.AspNetCore": "Warning"
-    },
-    "WriteTo": [
-      {
-        "Name": "Console",
-        "Args": {
-          "outputTemplate": "[{Timestamp:HH:mm:ss} {Level}] {SourceContext}{NewLine}{Message:lj}{NewLine}{Exception}{NewLine}"
-        }
+    "MinimumLevel": {
+      "Default": "Information",
+      "Override": {
+        "Microsoft": "Information",
+        "System": "Warning"
       }
-    ]
+    }
+  },
+  "Elasticsearch": {
+    "Uri": "http://localhost:9200"
   }
 }
